requirements.txt
requests
yara-python
tqdm
ssdeep
py-tlsh
jc-sdhash
flask